feat(ExercisesPanel): Extract `<SubPagesLinksPanel/>`
